Remove RefactoringASTParser from jdt.ls
- replace usage of jdt.ls RefactoringASTParser with
  the one in jdt.core.manipulation
